rename: 회원가입 URL 변경
